refactored index.js to utilize utility functions in order to outsource code
