+++ functions/handlers/users.js
@@ -1,0 +1,82 @@
+const { db } = require("../utility/admin")
+
+const config = require("../utility/config")
+
+const firebase = require("firebase")
+firebase.initializeApp(config)
+
+const {
+  validateSignupData,
+  validateLoginData
+} = require("../utility/validators")
+
+exports.signup = (req, res) => {
+  const newUser = {
+    email: req.body.email,
+    password: req.body.password,
+    confirmPassword: req.body.confirmPassword
+  }
+
+  const { valid, errors } = validateSignupData(newUser)
+
+  if (!valid) return res.status(400).json(errors)
+
+  let token, userId
+  firebase
+    .auth()
+    .createUserWithEmailAndPassword(newUser.email, newUser.password)
+    .then(data => {
+      userId = data.user.uid
+      return data.user.getIdToken()
+    })
+    .then(idToken => {
+      token = idToken
+      const userCredentials = {
+        email: newUser.email,
+        createdAt: new Date().toISOString(),
+        userId
+      }
+      return db.doc(`/users/${userId}`).set(userCredentials)
+    })
+    .then(() => {
+      return res.status(201).json({ token })
+    })
+    .catch(err => {
+      console.error(err)
+      if (err.code === "auth/email-already-in-use") {
+        return res.status(400).json({ email: `email is already in use` })
+      } else {
+        return res.status(500).json({ error: err.code })
+      }
+    })
+}
+
+exports.login = (req, res) => {
+  const user = {
+    email: req.body.email,
+    password: req.body.password
+  }
+
+  const { valid, errors } = validateLoginData(user)
+
+  if (!valid) return res.status(400).json(errors)
+
+  firebase
+    .auth()
+    .signInWithEmailAndPassword(user.email, user.password)
+    .then(data => {
+      return data.user.getIdToken()
+    })
+    .then(token => {
+      return res.json({ token })
+    })
+    .catch(err => {
+      console.log(err)
+      if (err.code === "auth/wrong-password") {
+        return res
+          .status(403)
+          .json({ general: "wrong credentials please try again" })
+      }
+      return res.status(500).json({ error: err.code })
+    })
+}
