+++ functions/utility/admin.js
@@ -1,0 +1,11 @@
+const admin = require("firebase-admin")
+
+var serviceAccount = require("../../functions/admin.json")
+admin.initializeApp({
+  credential: admin.credential.cert(serviceAccount),
+  databaseURL: "https://jobtrackerback.firebaseio.com"
+})
+
+const db = admin.firestore()
+
+module.exports = { admin, db }
