+++ functions/utility/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  apiKey: "AIzaSyDA52rltHRk76oU2HEMEyYqm4Y0vGf6sd0",
+  authDomain: "jobtrackerback.firebaseapp.com",
+  databaseURL: "https://jobtrackerback.firebaseio.com",
+  projectId: "jobtrackerback",
+  storageBucket: "jobtrackerback.appspot.com",
+  messagingSenderId: "643549026160",
+  appId: "1:643549026160:web:335b1600b4a11549ba21b5",
+  measurementId: "G-E1B73KCWM7"
+}
